Split up react and media component helpers
