+++ spec/helpers/media_component_helper_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe ReactHelper do
+describe MediaComponentHelper do
   describe 'render_video_component' do
     let(:media) { Fabricate(:media_attachment, type: :video, status: Fabricate(:status)) }
     let(:result) { helper.render_video_component(media.status) }
@@ -75,40 +75,6 @@ describe ReactHelper do
 
     it 'returns the correct react component markup' do
       expect(parsed_html.div['data-component']).to eq('Poll')
-    end
-  end
-
-  describe 'react_component' do
-    context 'with no block passed in' do
-      let(:result) { helper.react_component('name', { one: :two }) }
-
-      it 'returns a tag with data attributes' do
-        expect(parsed_html.div['data-component']).to eq('Name')
-        expect(parsed_html.div['data-props']).to eq('{"one":"two"}')
-      end
-    end
-
-    context 'with a block passed in' do
-      let(:result) do
-        helper.react_component('name', { one: :two }) do
-          helper.content_tag(:nav, 'ok')
-        end
-      end
-
-      it 'returns a tag with data attributes' do
-        expect(parsed_html.div['data-component']).to eq('Name')
-        expect(parsed_html.div['data-props']).to eq('{"one":"two"}')
-        expect(parsed_html.div.nav.content).to eq('ok')
-      end
-    end
-  end
-
-  describe 'react_admin_component' do
-    let(:result) { helper.react_admin_component('name', { one: :two }) }
-
-    it 'returns a tag with data attributes' do
-      expect(parsed_html.div['data-admin-component']).to eq('Name')
-      expect(parsed_html.div['data-props']).to eq('{"locale":"en","one":"two"}')
     end
   end
